Refactor UserAuditPage so that it's not as bloated
